Overwrite the cutoff with the WSR.

+++ src/simd/algorithm.hpp
@@ -36,7 +36,7 @@ namespace qmcplusplus {
       }
 
     template<typename T1, typename T2>
-    inline T2 accumulate_n(const T1* restrict in, int n, T2 res)
+    inline T2 accumulate_n(const T1* restrict in, size_t n, T2 res)
       {
 #pragma omp simd reduction(+:res)
         for(int i=0; i<n; ++i)
